Removed weight because powerups will not have weight

+++ Assets/Scripts/Utilities/ISpawnableObjects.cs
@@ -6,8 +6,8 @@ namespace Dots.Utils.Spawnable
     {
         float Speed { get; set; }
         Vector2 Direction { get; set; }
-        int RandX { get; set; }
-        int RandY { get; set; }
+        float RandX { get; set; }
+        float RandY { get; set; }
 
         void SetSpeedAndDirection();
         void BehaveWhenInteractWithBounds();
